Backport GitHub Actions integration to stable-4.11
The tweak in src/gasman.c resolves a compiler warning that'd cause
a CI failure (as we compile with -Werror)
