+++ .appveyor.yml
@@ -41,7 +41,7 @@ build_script:
         Where-Object pullRequestId -eq $env:APPVEYOR_PULL_REQUEST_NUMBER)[0].buildNumber) { `
         throw "There are newer queued builds for this pull request, failing early." }
   - SET "PATH=%CYG_ROOT%\bin;%PATH%"
-  - python -m pip install gcovr==4.1 # for coverage reporting later on
+  - python -m pip install gcovr==4.2 # for coverage reporting later on
   - bash -lc "gcc --version"
   - bash -lc "cd $APPVEYOR_BUILD_FOLDER && ./etc/ci-prepare.sh"
