+++ etc/ci-coveralls-merge.py
@@ -4,12 +4,18 @@
 
 from __future__ import print_function
 
+import hashlib
 import json
 import os
 import os.path
 import sys
 
-coverage_files = [ 'c-coveralls.json' ]
+def md5(fname):
+    hash_md5 = hashlib.md5()
+    with open(fname, "rb") as f:
+        for chunk in iter(lambda: f.read(4096), b""):
+            hash_md5.update(chunk)
+    return hash_md5.hexdigest()
 
 print("Merging coveralls json coverage") 
 
@@ -30,15 +36,24 @@ if not 'source_files' in merged:
     print("warning: source_files not in JSON dictionary read from gap-coveralls.json")
     merged['source_files'] = []
 
-for fn in coverage_files:
-    print("file: %s" % (fn,), end="")
-    if os.path.isfile(fn):
-        with open(fn, 'r') as f:
-            cover = json.load(f)
-            merged['source_files'].extend(cover['source_files'])
-        print(" done.")
-    else:
-        print(" not found.")
+print("merging c-coveralls.json")
+with open('c-coveralls.json', 'r') as f:
+    input = json.load(f)
+    sf = merged['source_files']
+    for file in input['files']:
+        name = file["file"]
+        print("processing ", name)
+        # TODO: normalize filename, remove leading "./"?
+        source_digest = md5(name)
+        max_line_number = max([line["line_number"] for line in file["lines"]])
+        coverage = [None] * max_line_number
+        result = { "name" : name, "source_digest" : source_digest, "coverage" : coverage }
+        print("  MD5 hash = ", source_digest)
+        for line in file["lines"]:
+            if line["gcovr/noncode"]:
+                continue
+            coverage[line["line_number"]-1] = line["count"]
+        sf.append(result)
 
 if len(merged['source_files']) > 0:
     print("Merged:")
